Coding Assignment 4/requirements.txt
openai
requests
beautifulsoup4
lxml
pandas
supabase
streamlit
modal
python-dotenv
pypdf
pdfplumber